{"schema":"decodex/commit/1","summary":"Modularize capture-frame rendering","authority":"manual"}
